+++ torchutils/__init__.py
@@ -3,4 +3,4 @@ import torchutils.models
 import torchutils.trainer
 import torchutils.logging
 
-__version__ = torchutils.utils.Version(1, 0, 0)
+__version__ = torchutils.utils.Version('1.1.0')