+++ torchutils/logging/base.py
@@ -1,4 +1,4 @@
-from typing import overload
+from typing import overload, List
 from abc import ABC, abstractmethod
 
 
@@ -11,11 +11,23 @@ class TrainerLogger(ABC):
         ...
 
     @abstractmethod
-    def log(self, **kwargs):
+    def log_score(self, **kwargs):
         ...
 
     @abstractmethod
-    def _flush(self):
+    def log_info(self, msg):
+        ...
+    
+    @abstractmethod
+    def log_error(self, msg):
+        ...
+
+    @abstractmethod
+    def _flush_step(self):
+        ...
+
+    @abstractmethod
+    def _flush_epoch(self):
         ...
 
     @abstractmethod
@@ -24,19 +36,38 @@ class TrainerLogger(ABC):
 
 
 class LoggingHandler(ABC):
-    def init(self):
-        ...
+    def __init__(self, loggers=[]):
+        self._loggers: List[TrainerLogger] = []
+        self.add_loggers(loggers)
 
-    def log(self):
-        ...
+    def add_logger(self, logger):
+        if isinstance(logger, trainerlogger):
+            self._loggers.append(logger)
+
+    def remove_logger(self, logger):
+        self._loggers.remove(logger)
+
+    def add_loggers(self, loggers: List[TrainerLogger]):
+        for logger in loggers:
+            if isinstance(logger, trainerlogger):
+                self._loggers.extend(loggers)
+
+    def clear_loggers(self):
+        self._loggers.clear()
+
+    def init(self):
+        for logger in self._loggers:
+            logger.open()
 
     def step(self):
-        ...
+        for logger in self._loggers:
+            logger._flush_step()
 
-    def close(self):
-        ...
+    def epoch(self):
+        for logger in self._loggers:
+            logger._flush_epoch()
 
-
-
-
+    def end(self):
+        for logger in self._loggers:
+            logger.close()