+++ torchutils/metrics/base.py
@@ -37,43 +37,62 @@ class TrainerMetric(ABC):
 
 
 class MetricHandler(object):
-    _scores_by_name: Mapping[Str, TrainerMetric] = OrderedDict()
-    _metric_by_name: Mapping[Str, TrainerMetric] = OrderedDict()
+    _scores: Dict[Str, TrainerMetric] = OrderedDict()
+    _metrics: Dict[Str, TrainerMetric] = OrderedDict()
 
     def __init__(self, *score_names):
-        self._scores_by_name: Mapping[Str, TrainerMetric] = OrderedDict()
-        self._metric_by_name: Mapping[Str, TrainerMetric] = OrderedDict()
-        all_score_names = self.__class__._scores_by_name
-        for score_name in score_names:
-            if score_name not in self.__class__._scores_by_name.keys():
-                warnings.warn(f"Score {score_name} is not registered", RuntimeWarning)
-            else:
-                trainer_metric = self.__class__._scores_by_name[score_name]
-                self._scores_by_name[score_name] = trainer_metric 
-                self._metric_by_name[trainer_metric.__class__.__name__] = trainer_metric
+        self._scores: Dict[Str, TrainerMetric] = OrderedDict()
+        self._metrics: Dict[Str, TrainerMetric] = OrderedDict()
+        # TODO: why it is here? all_score_names = self.__class__._scores.copy()
+        self.add_scores(score_names)
     
+    def add_scores(self, *score_names: Str):
+        for score_name in score_names:
+            if score_name not in self.__class__._scores.keys():
+                warnings.warn(f"Score {score_name} is not registered", 
+                        RuntimeWarning)
+                continue
+            trainer_metric = self.__class__._scores[score_name]
+            self._scores[score_name] = trainer_metric 
+            self._metrics[trainer_metric.__class__.__name__] = trainer_metric
+
+    def remove_scores(self, *score_names: Str):
+        for score_name in score_names:
+            if score_name not in self.__class__._scores.keys():
+                warnings.warn(f"Score {score_name} is not registered", 
+                        RuntimeWarning)
+                continue
+            metric = self._scores.pop(score_name)
+            if metric not in self._scores.values():
+                self._metrics.pop(metric.__class__.__name__)
+
+    def clear_scores(self):
+        self._metrics.clear()
+        self._scores.clear()
+
+    #TODO: write register_scores
     @classmethod
     def register_metric(cls, trainer_metric):
         for score_name in trainer_metric.score_names:
-            cls._scores_by_name[score_name] = trainer_metric 
-            cls._metric_by_name[trainer_metric.__class__.__name__] = trainer_metric
+            cls._scores[score_name] = trainer_metric 
+            cls._metrics[trainer_metric.__class__.__name__] = trainer_metric
 
     #TODO: create class methods
     def get_score_names(self) -> Iterable:
-        return iter(self._scores_by_name.keys())
+        return iter(self._scores.keys())
 
     def get_metric_names(self) -> Iterable:
-        return iter(self._metric_by_name.keys())
+        return iter(self._metrics.keys())
 
     def get_metric_instance(self, metric_name):
-        return self._metric_by_name[metric_name]
+        return self._metrics[metric_name]
 
     def get_score_values(self, *score_names) -> Mapping[Str, Any]:
         if len(score_names) == 0: score_names = self.get_score_names()
-        return { score_name: self._scores_by_name[score_name] \
+        return { score_name: self._scores[score_name] \
             .get_scores(score_name) for score_name in score_names }
 
     def set_scores_values(self, x, y, y_pred):
         for metric_name in self.get_metric_names():
-            self._metric_by_name[metric_name].set_scores(x, y, y_pred)
+            self._metrics[metric_name].set_scores(x, y, y_pred)