+++ torchutils/utils/__init__.py
@@ -6,6 +6,8 @@ from .mthdutils import hybridmethod
 from .decorators import verbose, profile
 from .config import INIObject
 from .hash import Hashable
+
+
 import torchutils.utils.config
 #import torchutils.data.utils.preprocessing
 import numpy as np